fix(assistant): don't replay tool turns; clear-history return type; svg favicon
Three fixes shipping together because they're all the same iteration of
"the assistant works end-to-end now."

  - Persistence: stop saving the empty-content "assistant" placeholder
    and the "tool" result rows that bracket each tool call. Replaying
    them on the next request put the conversation in a state Gemini
    rejects with `INVALID_ARGUMENT: function response turn comes
    immediately after a function call turn` — Anthropic and Ollama
    happen to tolerate it. The model can re-fetch via tools whenever it
    needs the same data.

  - loadRecentHistory: defensively skips empty-content rows and any
    role other than user/assistant, so leftover broken pairs from
    older builds don't keep poisoning new conversations.

  - clear-chat 500: ChatMessageRepository.deleteByUserId returned
    `long`, but Spring Data JPA's @Modifying queries only accept
    void / int / Integer. Switched to int (and propagated through
    AssistantService.clearHistory + AssistantController.clear). The
    DELETE /api/assistant/history endpoint now succeeds.

  - app/icon.svg + app/apple-icon.svg: brand-red rounded-square with
    the Sparkle glyph and a white center dot. Next.js App Router
    auto-discovers these and emits the correct <link rel="icon"> +
    apple-touch-icon. No manifest needed.

`./mvnw test` still 9/9 green; frontend lint / typecheck / vitest
clean.

+++ services/api/src/main/java/com/dsofikitis/chronos/ai/AssistantController.java
@@ -32,7 +32,7 @@ public class AssistantController {
 
     @DeleteMapping("/history")
     public Map<String, Object> clear(@CurrentUserId UUID userId) {
-        long deleted = assistant.clearHistory(userId);
+        int deleted = assistant.clearHistory(userId);
         return Map.of("deleted", deleted);
     }
